Remove default value for property YubiKey.piv
This fixes a future TypeError in `PivManager`, currently hidden by two
mistakes that cancel each other.

PivManager checks if `YubiKey.piv` is truthy, and sends
`YubiKey.piv.certificates` to `PivCertificates` if it is. However, the
intent of setting the default value `piv: {}` was to set it to default
to an empty object, which is a truthy value. This mistake was not
discovered because `piv: {}` actually evaluates to falsy because `{}` is
interpreted as a code block returning nothing. @dagheyman discovered
that Qt Creator was suggesting changing this to `({})`, which does
evaluate to an empty object and exposes the bug.

Removing the default value makes the truthiness check in `PivManager` do
the right thing.